Moving auth service out of index.
Next is to create an interface.js file.
